implement prop types, move Launch related components into /components/Launch
