+++ package.json
@@ -24,6 +24,7 @@
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@fortawesome/react-fontawesome": "^0.1.14",
     "normalize.css": "^7.0.0",
+    "prop-types": "^15.7.2",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
     "react-hot-loader": "^4.12.14",
@@ -91,6 +92,7 @@
     "node-sass": "^4.0.0",
     "postcss-loader": "^2.0.9",
     "prettier": "^1.14.3",
+    "prop-types": "^15.7.2",
     "react-test-renderer": "^16.2.0",
     "sass-loader": "^6.0.6",
     "sass-resources-loader": "1.3.2",
