+++ styles/_style.sass
@@ -34,7 +34,7 @@ body
           margin: 0
 
           > li
-            margin: 4px;
+            margin: 4px
             padding: 4px
             background: #ccc
             border-bottom: 2px solid #ccc
